uploader/cloud_function/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-storage==2.*
google-auth==2.*
flask>=2.0,<4.0
